requirements.txt
# MVidarr Enhanced - Production Dependencies

# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
alembic==1.13.1
mysqlclient>=2.2.0

# HTTP and API
requests==2.31.0
urllib3==2.1.0
httpx==0.25.2

# Image Processing
Pillow==10.1.0
opencv-python-headless>=4.8.0

# Media Processing
python-ffmpeg==2.0.12
moviepy==1.0.3

# Data Processing
python-dateutil==2.8.2
python-dotenv==1.0.0
pydantic==2.5.0
marshmallow==3.20.1

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Security
bcrypt==4.1.2
cryptography>=42.0.0
PyJWT==2.8.0

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
psutil==5.9.6
schedule==1.2.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Additional utilities
yarl==1.9.4
aiofiles==23.2.1
python-slugify==8.0.1
bleach==6.1.0
humanize==4.9.0

# Video processing (legacy support)
yt-dlp>=2024.1.0